+++ CHANGELOG.md
@@ -1,17 +1,26 @@
 # svelte-pathfinder changelog
 
+# 2.2.0
+
+* Use single export instead of multiple.
+* Fix `prefs` export which was missed in PR.
+
 # 2.1.2
+
 * Fix boolean values convertation.
 
 # 2.1.1
+
 * Minor fix of new query array param formating.
 
 # 2.1.0
+
 * Improve query string parsing
 * Support different formats of query string array processing, eg. `?foo[]=1&foo[]=2` (default) and NEW alternative way `?foo=1,2`.
 * Support arrays in `path` eg. param`/:variants` in example `/foo|bar|baz` will be parsed as array `$path.variants === ['foo', 'bar', 'baz']`.
 
 # 2.0.0
+
 * Re-write to Typescript (first edition).
 * Add `svelte` to devDeps to get typings.
 * Improve query string parsing, better support for nesting objects.